requirements.txt
aiohttp==3.9.5
pycryptodome==3.20.0